requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.1
requests==2.28.1
holidays==0.54